yes text and
